Plot time line with images and descriptions of each event

The time line is plotted horizontally with ticks for years. The labels with images and descriptions are above and/or below that line.

plot_timeline(
  events_df,
  ys,
  width = 100,
  description_width = 20,
  expand_x = c(0.1, 0.1),
  expand_y = c(0.05, 0.05)
)

Arguments

events_df
A data frame with at least these columns:

image

Paths to the images to use in the label. NA for items without an image.

date_published

A vector of Date objects for dates when the event of interest was published. Note that the actual time when those events occurred is most likely earlier, sometimes much earlier than the publication date, and the events might have become quite influential before their publication. But the publication date is the easiest way to get an actual date.

description

Short descriptions of the events. The plot won't look good if the descriptions are too long.

ys
A numeric vector of the y coordinates of the items. Since I don't know how to implement the ggrepel algorithm to make sure that the labels don't overlap for geom_richtext, I have to manually set the y coordinates to make sure that the labels don't overlap and look good.

width
Width of the labels in pixels. How the labels will look depends on the size of the plot. Would be cool if this can be specified in terms of units within the plot.

description_width
Width of the description text in characters to wrap the text in the labels.

expand_x
A numeric vector of length 2 of the proportion to expand the x axis on the left and the right. This is a way to manually make sure that the labels are not cropped off at the edge of the plot.

expand_y
Same as expand_x, but for the y axis.

Value

A ggplot2 object for the plot.

Details

These are the sources of images that are not stated to be under public domain found online with filter "free to share and use". No changes were made to the images unless indicated. The author and license, if found, are listed here as well.

drosophila.jpg

http://gompel.org/images-2/drosophilidae

zebrafish.jpg

https://thumbs.dreamstime.com/m/zebrafish-zebra-barb-danio-rerio-freshwater-aquarium-fish-isolated-white-background-50201849.jpg

ciona.jpg

http://www.habitas.org.uk/marinelife/tunicata/cioints.jpg

xenopus.jpg

https://en.wikipedia.org/wiki/African_clawed_frog#/media/File:Xenopus_laevis_02.jpg by Brian Gratwicke. License: https://creativecommons.org/licenses/by/2.0/

celegans.jpg

https://en.wikipedia.org/wiki/Caenorhabditis_elegans#/media/File:Adult_Caenorhabditis_elegans.jpg by Kbradnam at English Wikipedia. License: https://creativecommons.org/licenses/by-sa/2.5/ A smaller version of the original is used here.

arabidopsis.jpg

http://parts.igem.org/wiki/images/b/bd/Plants_Arabidopsis_thaliana_400px.jpg

skull.jpg

http://pngimg.com/download/42558 License: https://creativecommons.org/licenses/by-nc/4.0/ The original was compressed and converted to jpg here.

platynereis.jpg

https://en.wikipedia.org/wiki/Epitoky#/media/File:PlatynereisDumeriliiFemaleEpitoke.tif By Martin Gühmann. A smaller jpg version of the original is used here. License: https://creativecommons.org/licenses/by-sa/4.0/

yeast.jpg

https://en.wikipedia.org/wiki/Shmoo#/media/File:Shmoos_s_cerevisiae.jpg By Pilarbini. This is a smaller version of the original. License: https://creativecommons.org/licenses/by-sa/4.0/deed.en