Number of publications per year

Plot bar plot of the number of publications per year. Preprints are excluded as the dates are incoherent with those for published papers. I find facetting makes the plot easier to read than filling with different colors.

pubs_per_year(
  pubs,
  facet_by = NULL,
  binwidth = 365,
  preprints = FALSE,
  n_top = Inf,
  sort_by = c("first_appeared", "count", "recent_count")
)

Arguments

pubs
A data frame with at least these columns:

journal

Name of the journal of the paper.

year

Year when the paper was published.

There must be one row per publication or per method or species for each title if faceting by those. If facetting, then a column whose name is the value in `facet_by` must be present.

facet_by
Name of a column for facetting.

binwidth
Width of bins for the histogram in days.

preprints
Logical, whether preprints should be included.

n_top
Number of categories with the most publications to plot in facets; the other categories are lumped into "other".

sort_by
How to sort the facets. first_appeared means the category that appeared earlier will be nearer to the top. count means the category with more count (number of publications) will be nearer to the top. Ignored if not facetting.

Value

A ggplot2 object.